fixed all windows warnings
Former-commit-id: e6419d4790defc9ab4c4f1b2a333b7ce51267de9
Former-commit-id: 18ab1bf0997706e03de721808904fa6a32a92a47

+++ src/architecture/utilitiesSelfCheck/gaussMarkovCheck.cpp
@@ -81,7 +81,7 @@ uint64_t testGaussMarkov()
     Eigen::Vector2d minOut;
     minOut.fill(0.0);
     
-    numPts = 1e6;
+    numPts = (int64_t) 1e6;
     
     noiseOut.resize(2,numPts);
